+++ autode/path/__init__.py
@@ -1,0 +1,5 @@
+from autode.path.path import Path
+from autode.path.adaptive import AdaptivePath
+
+__all__ = ['Path',
+           'AdaptivePath']
