+++ doc/reference/bond_lengths.rst
@@ -2,7 +2,7 @@
 Bonds
 *****
 
-.. automodule:: autode.bond_lengths
+.. automodule:: autode.bonds
    :members:
    :undoc-members:
    :special-members: __init__
