+++ setup.py
@@ -11,6 +11,7 @@ setup(name='autode',
       packages=['autode',
                 'autode.conformers',
                 'autode.pes',
+                'autode.path',
                 'autode.neb',
                 'autode.reactions',
                 'autode.smiles',
